enabling functionality test to run under Wondows (via git-bash)

+++ test/functional_tests/p6/runp6.sh
@@ -4,6 +4,12 @@ wrkDir=`pwd`
 logFile="p6.log"
 paramFile="p6.params"
 
+if [[ "$(uname)" == CYGWIN* ]] || [[ "$(uname)" == MINGW* ]] || [[ "$(uname)" == MSYS* ]]
+then
+    echo "Test SKIPPED on Windows"
+    exit 0
+fi
+
 mv data/* "$wrkDir"
 rm -rf data
